src/hackathon-18march-boilerplate/provider/requirements.txt
orca-platform-sdk-ui
uvicorn
fastapi
agno
httpx
openai
